Updated AL-Go System Files (#5)
## Preview

Note that when using the preview version of AL-Go for GitHub, you need
to Update your AL-Go system files, as soon as possible when told to do
so.

### Issues

Issue #446 Wrong NewLine character in Release Notes
Issue #453 DeliverToStorage - override fails reading secrets
Issue #434 Use gh auth token to get authentication token instead of gh
auth status
Issue #501 The Create New App action will now use 22.0.0.0 as default
application reference and include NoImplicitwith feature.


### New behavior

The following workflows:

- Create New App
- Create New Test App
- Create New Performance Test App
- Increment Version Number
- Add Existing App
- Create Online Development Environment

All these actions now uses the selected branch in the **Run workflow**
dialog as the target for the Pull Request or Direct COMMIT.

### New Settings

- `UseCompilerFolder`: Setting useCompilerFolder to true causes your
pipelines to use containerless compiling. Unless you also set
`doNotPublishApps` to true, setting useCompilerFolder to true won't give
you any performance advantage, since AL-Go for GitHub will still need to
create a container in order to publish and test the apps. In the future,
publishing and testing will be split from building and there will be
other options for getting an instance of Business Central for publishing
and testing.
- `vsixFile`: vsixFile should be a direct download URL to the version of
the AL Language extension you want to use for building the project or
repo. By default, AL-Go will use the AL Language extension that comes
with the Business Central Artifacts.

### New Actions

- **DetermineArtifactUrl** is used to determine which artifacts to use
for building a project in CI/CD, PullRequestHandler, Current, NextMinor
and NextMajor workflows.

### License File

With the changes to the CRONUS license in Business Central version 22,
that license can in most cases be used as a developer license for
AppSource Apps and it is no longer mandatory to specify a license file
in AppSource App repositories.
Obviously, if you build and test your app for Business Central versions
prior to 21, it will fail if you don't specify a licenseFileUrl secret.

Co-authored-by: aholstrup1 <aholstrup1@users.noreply.github.com>