[main] Update baseline package handling (#21)
Update baseline package handling to not depend on BCArtifacts anymore. 

CLEAN mode should grab the latest version out of main as the baseline 
Default/Translated mode will use the package defined in Packages.Json as
the baseline. This should always be a build out of release/22

Here's a test run of the CICD workflow out of this branch:
https://github.com/microsoft/BCApps/actions/runs/5001796789

---------

Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>