[Copilot] Marketing Text Adjustments (#256)
This PR brings a few fixes to the Marketing Text and AI module.

AI SDK:
- `CurrPage.Update()` is passed `false` in the AI SDK, fixing a bug with
unwanted changes saved to the copilot capabilities

Marketing Text:
- `CanSuggest()` and `IsEnabled()` checks are modified to correctly
manage visibility and actions for Entity/Marketing Text. This is also
reflected in the `GenerateSuggestion`, to ensure that all required
checks pass before a call to AI service is allowed.