Make sure feature telemetry state updates can be done in quick succession (#240)
**Problem:**
If feature telemetry state update calls are done quicker than it takes
to start a session (a few milliseconds), then the update might not go
through correctly.

**Solution:**
If the same feature telemetry instance needs to update the state of a
feature more than once in a row, then make sure that the previous
session had time to complete.

---------

Co-authored-by: Ihor Handziuk <ihhandzi@microsoft.com>