Introduce "User Details" module (#254)
**Problem:**
There is a lot of information that would be valuable to see on the
`Users` page in a list view for search and filtering purposes (such as
details about user plans, permission sets, user properties etc). But
it's not possible to have table extensions for system tables, so such
fields couldn't have been added directly.

**Solution:**
- Introduce a new temporary table `User Details` that fetches
information from the `User` table and other related tables.
- Introduce a new page that surfaces the information from the `User
Details` table in a list view.

**Note:**
The new page will be referenced in the `Navigate` pane of the `Users`
page.

---------

Co-authored-by: Ihor Handziuk <ihhandzi@microsoft.com>