Update workitem validation logic (#365)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

Updated the workitem validation such that
* All PRs from forks must reference a GitHub issue 
* All PRs must reference an ADO workitem (regardless of whether they're
from a fork or not)

To help link BCIdea workitems to pull requests, I've added a stage to
the WorkitemValidation.yaml that looks for ADO workitems in the linked
issues and automatically links them to the pull request.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes AB#476348

---------

Co-authored-by: aholstrup1@microsoft.com <aholstrup1@microsoft.com@users.noreply.github.com>
Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>