[Email] Use Sent From email in EmailViewer (#380)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
The EmailViewer is using the Email Account's "Email Address" field. This
is always the current user's email and not the email address of the user
who sent the email.

This PR uses the stored Sent Email's "Sent From" to show the correct
email address of the sender.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes AB#491830