Add verify execute permissions (#379)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
This PR addresses an inconsistency between our internal validation and
the validation on BCApps. On BCApps, we do not check if objects have the
appropriate permissions. This PR introduces that check.


#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes AB#492320

---------

Co-authored-by: aholstrup1@microsoft.com <aholstrup1@microsoft.com@users.noreply.github.com>