[Copilot] Hide status field when data movement is not enabled (#439)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
**Problem** 
It is currently misleading that copilots are active/enabled when the
data movement toggle is off.
If the toggle is off and copilots show active, the copilots are actually
deactivated due to data movement not being allowed.

**Solution**
Hide the Status column whenever data movement toggle is off.


![image](https://github.com/microsoft/BCApps/assets/18476646/580e305c-f12f-4e3f-b13a-ad78c9f50254)

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes AB#491279