Add a script to create local development environment (#417)
#### Summary <!-- Provide a general summary of your changes -->
Create a script to create local development environment for working with
apps from the repository.

The NewDevEnv.ps1 script creates a docker container (if it doesn't
exist), builds the apps, based on the provided parameters, and publishes
the apps to the container.

Test run:
https://github.com/mazhelez/BCApps/actions/runs/7225774768/job/19689897656

Note:
- The created container contains only apps from the BCApps repository.

To be considered for future enhancement:
- The created container contains all BC apps (by Microsoft): the ones
from this repo are rebuilt, the rest are taken from the BC artifact.
- Rebuild apps functionality. Now the rebuild happens only after manual
clean up the cache folder.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes AB#492033

---------

Co-authored-by: Alexander Holstrup <117829001+aholstrup1@users.noreply.github.com>