#398 GetSeed for CU Any (#402)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
GetSeed() procedure for CU Any. It required a small refactoring for the
SetDefaultSeed() procedure to explicitly define the seed instead of the
implicit definition using Randomize(). The value used as seed is still
the same.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #398

<br /><br />Internal work item: AB#493044