Add procedures with defaultButton true (#440)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

In most cases where we want the user to confirm an action, we set the
default button to true. This will set the focus to the specific button
so that the user can simply press enter and continue.

DefaultButton should be set to false on critical actions manually.

- [ ] Additional tests should be added

*Note*: When I first read the DefaultButton parameter, I thought that
this parameter would also be applied if the confirmation page was
canceled by pressing <kbd>Esc</kbd>, for example. However, this is not
the case. The DefaultButton parameter only sets the focus on the button.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #436





<br /><br />Internal work item: AB#495008