Turn AW0009 and AS0059 warnings into errors (#463)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Turn codecop AW0009 and AS0059 into errors.

Using pragmas instead of fixing due to following reasons:
AW0009: Table needs to be changed to temporary. Codecop was added due to
their specific apparently.

AS0059: The places where the bitmaps are used, aren't one of the
supported places for media/mediaset fields.
Fx, list of records. The only time it works, is if it is showing in a
tile view, which for Emails, we are not.
The MediaCleanup uses Tenant Media record directly, which is a blob.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#487172](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/487172)