Remove the faulty permission check in the RemoveOrphanedEntraGroups function (#470)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
**Problem:**
The code for deleting orphaned security groups starts with:
```
if not User.WritePermission() then
    exit;
```
But in SaaS no entitlement has insert permission to the user table, so
this check always fails. Unfortunately, there is no way to check for
delete permission separately.

**Solution:**
Remove the check and protect the deletion operation with `if not ...
then`.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#496007](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/496007)