[Retention Policy] Stop run once limit is hit (#496)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
**Problem**
The RP JQ does not stop running when the limit of records is hit. It
continues on to the next table. Whenever the limit is hit, a new JQ will
be created.

**Solution**
Set `EndCurrentRun` to true when the limit is hit. That stops the entire
RP from proceeding further until the next JQ run.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#497286](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/497286)