Fix the error message in the TestMixedPlansDetected test (#510)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
**Problem:**
The test `TestMixedPlansDetected` is currently not verifying the correct
error message. It was always meant to check for non-admin error, as the
test does not set up the current user as the admin user. The PR #497
uncovered this issue, but the test is disabled in the BC Apps gate, so
this was not cought before.

**Solution:**
Use the correct error message in the test verification step.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#497693](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/497693)