Introducing new Business Foundation layer - No. Series (#319)
This PR introduces the new Business Foundation layer. This layer adds
common Business logic on top of System Application, making it easier to
create extensions and ISV solutions.
We are extracting the No. Series functionality from Business Central in
a non-breaking solution. We are introducing two new interfaces, "No.
Series" and "No. Series - Batch".
"No. Series" is the general interface for interacting with No. Series,
however it may lock the No. Series table when No. Series are not set up
to use sequences. In cases where you need maximum performance you can
use the "No. Series - Batch". This interface allows you to work with No.
Series without calling the database every time and save the final state
to the database with SaveState.

Fixes
[AB#471519](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/471519).

---------

Signed-off-by: Gert Robyns <grobyns@microsoft.com>
Co-authored-by: Gert Robyns <grobyns@microsoft.com>
Co-authored-by: aholstrup1 <aholstrup1@microsoft.com@users.noreply.github.com>
Co-authored-by: Ihor Handziuk <ihor.handziuk@microsoft.com>