Add build configuration for strictmode (#502)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
The PR sets up the needed scripts for running builds in StrictMode. We
enable StrictMode by setting the StrictMode tag on a release branch.
Once the tag is set we pick the latest available bcartifact of a given
version as the baseline.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#497503](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/497503)

---------

Co-authored-by: aholstrup1 <aholstrup1@microsoft.com@users.noreply.github.com>
Co-authored-by: Maria Zhelezova <43066499+mazhelez@users.noreply.github.com>