No. Series: Add legacy test and fix bad date change (WorkDate) (#559)
Adding old legacy tests for No. Series Management to make sure the
changed code works with the old tests and fix an issue where we infer
the date the No. Series Line is modified is the WorkDate, when in fact
we don't know (the No. Series Management handles updating the date
properly, but only if this code does not modify it first).

Fixes
[AB#471519](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/471519)