Allow plus sign in field names when using Edit-In Excel (#597)
The issue

Edit-In Excel module does not externalize page field names the same way
the platform does, when the field name contains an plus sign.
For a page field defined as:
```
field("new vendor+s name"; rec."New Vendor+s Name")
{
	ApplicationArea = All;
}
```

The platform will externalize this name as `new_vendor_x002b_name` while
the Edit-In Excel module will externalize it as `new_vendor_name`.

Expected behavior

Within reason the platform and the Edit-In Excel module should
externalize fields the same way to avoid discrepancies between the API
metadata and the Edit-In Excel field bindings.

Steps to reproduce

Create a page field with an plus sign in its name on a card page, then
try to export its corresponding list page using Edit-In Excel and
observe the error dialog in the Excel side pane where add-ins are
loaded.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#498449](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/498449/)

Co-authored-by: Matti Nielsen <mattinielsen@microsoft.com>