Add technical version to feature key mgmt page (#598)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
We have changed the content of the feature key management system table
column "Mandatory by" and then we have added a new column to hold the
technical version as well. This PR is just to add the new column and
some views to the page.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#497816](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/497816/)

---------

Co-authored-by: Matti Nielsen <mattinielsen@microsoft.com>