[OAuth] Reintroduce OAuthControlAddin for OnPrem support (#621)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
OAuthControlAddin was obsoleted in previous PR in favor of
OAuthIntegration.

Both should exist, as OAuthIntegration is used to create an
authorization link and needs a min height and width to do that. The
OAuthControlAddin has no height and width.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#502131](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/502131/)