[24.0] Fix issues with PowerBIManagement and OAuthIntegration (#639)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

1. Add missing reference to powerbi.js in PowerBIManagement
2. Also reintroduce OAuthControlAddin as both should exist.
OAuthIntegration is used to create an authorization link and needs a min
height and width to do that. The OAuthControlAddin has no height and
width.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#502394](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/502394)