Add subscriber to system trigger OpenFeatureManagement (#658)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
A link to the feature management page has been added to the system menu
which calls a system trigger, this PR adds a subscriber to that event
which opens the Feature Management page.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#502096](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/502096/)

---------

Co-authored-by: Matti Nielsen <mattinielsen@microsoft.com>
Co-authored-by: Natalie Karolak <34504100+NKarolak@users.noreply.github.com>