Determine maxAllowedObsoleteVersion from config in main (#856)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Determine maxAllowedObsoleteVersion from config in main

Right now it's not possible to backport PRs with 25.0 as an obsolete tag
because we didn't update the config value when we branched:
https://github.com/microsoft/BCApps/actions/runs/8449290970?pr=841

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#523764](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/523764)