Move OpenXml type to framework (#917)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Breaking change in OpenXml3, where type OpenXmlUnknownElement is moved
to the assembly DocumentFormat.OpenXml.Framework. The
OpenXmlUnknownElement type reference has been removed in Base
application and the functionality is replaced with a function in the BC
OpenXml add-in which prevents compile issues with the new assembly.
#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#524860](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/524860)