Only table with type Normal is supported in Data Classification Worksheet and Data Privacy Utilities (#903)
#### Summary <!-- Provide a general summary of your changes -->

Add the public procedure `IsSupportedTable`: returns True if the table
has type Normal and not obsoleted.

Will need to uptake in NAV for "Data Privacy Utilities" after this is
merged. (see
[!180681](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_git/NAV/pullrequest/180681))

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#523733](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/523733)