Add GetCollectionType to Edit-In Excel API (#550)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

The issue
We have received this issue:
https://github.com/microsoft/ALAppExtensions/issues/21940 where the
programmer wants to move all of the filters from one field to another.
The only thing missing for this capability is that we should be able to
dynamically detect if the filter is a AND or OR expression, this is not
possible in the current implementation.

Expected behavior
We should expose the "Edit in Excel Filter Collection Type" property on
the filter through the API, such that clients of the API can dynamically
detect what collection type the filter has.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#498510](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/498510/)

---------

Co-authored-by: Matti Nielsen <mattinielsen@microsoft.com>