[AzureADTenant] Ensure Entra tenant ID is formatted consistently in helper function (#684)
#### Summary
The Entra tenant ID returned from the System Application module is
picked from a server setting, and is then immediately returned to the
caller. There is no validation on its format.
For backwards compatibility, this will still be the case, but if the
setting is a GUID (which should be the case in most supported cases), it
will be formatted consistently. This makes our code using this function
more robust.

#### Work Item(s)
Fixes
[AB#504179](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/504179)