Cleanup the Edit-In Excel module docs (#1003)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
We merged in a PR updating the Edit-In Excel module API
https://github.com/microsoft/BCApps/pull/550, but forgot about updating
the documentation, namely the README.md which had examples using the old
API.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes
[AB#527688](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/527688/)

Co-authored-by: Matti Nielsen <mattinielsen@microsoft.com>