Fix error in AFS shared key authorization (#961) (#1006)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->

This is a PR, that should fix the error that occured when creating or
sending file to Azure File Share with AFS Library.
I also created a simple test to ensure it doesn't happen again.

The issue was with two things:
1. Sorting of the headers - In BC the order of the headers in the string
to sign was different than in the documentation, which led azure to
calculate a different hash for the authorization. I added sorting of the
headers in the same way as implemented in ABS module, which should
ensure correct order.
2. The header x-ms-content-length required when using CreateFile call,
was added to the content headers, which led to it not being correctly
added to the string to sign. I changed it to be added to request
headers.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #961 #685 


Fixes
[AB#532987](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/532987)