[Main] One Line Change: remove page control of drag and drop (#1023)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Remove page control of drag and drop, because client team decide to
postpone this one from 25.x to 26.x.

For now the compiler supports it, but it will turn to an error in the
future. So remove it. Add them back when supported.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->

[AB#484133](https://dynamicssmb2.visualstudio.com/Dynamics%20SMB/_workitems/edit/484133/)

Co-authored-by: wenjiefan <wenjiefan@microsoft.com>