Refactor FindBlobsByTags method to accept OptionalParameters (#1064)
<!-- Thank you for submitting a Pull Request. If you're new to
contributing to BCApps please read our pull request guideline below
* https://github.com/microsoft/BCApps/Contributing.md
-->
#### Summary <!-- Provide a general summary of your changes -->
Add `OptionalParameters` to `FindBlobsByTags` to provide the possibility
of setting `marker` and `maxresults`. This is required to use result
paging.

#### Work Item(s) <!-- Add the issue number here after the #. The issue
needs to be open and approved. Submitting PRs with no linked issues or
unapproved issues is highly discouraged. -->
Fixes #893 

Fixes
[AB#524630](https://dynamicssmb2.visualstudio.com/1fcb79e7-ab07-432a-a3c6-6cf5a88ba4a5/_workitems/edit/524630)

Co-authored-by: Christoph Blank <christoph.blank@kumavision.com>